# Little Optimization for RoPE Computation